quantedge/requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
rich>=13.0.0
schedule>=1.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
groq>=0.11.0
anthropic>=0.39.0
python-dotenv>=1.0.1
